Adding logging to assistant
This commit contains major alterations that add support for logging information in HyperDx.
As of now, we are going to keep track of the conversation between the user and the assistant by logging the messages and the actions call.
We are also tracking latency.
There were some refactorings in the code, such as the creation of the utils/ folder to store code common to all the assistants actions.
This commit also requires the update of the client-side.